bot/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==22.2.0
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.1.0
greenlet==2.0.2
idna==3.4
mojang==0.1.9
multidict==6.0.4
Pillow==8.3.2
python-dotenv==0.20.0
PyYAML==5.4.1
requests==2.28.2
SQLAlchemy==2.0.9
typing-extensions==4.5.0
urllib3==1.26.15
whois==0.9.27
yarl==1.8.2

git+https://github.com/Rapptz/discord.py@master#egg=discord.py
git+https://github.com/Rapptz/discord-ext-menus@master#egg=discord-ext-menus
git+https://github.com/Gelbpunkt/aiowiki
